Alpha 1.0.46 - Continuando a reorganização de diretórios nos arquivos frontend
