+++ Frontend/.env
@@ -1,5 +1,5 @@
 VITE_DEV = "Y"
 VITE_APP_NOME = "Módulo de Manutenções"
-VITE_APP_VERSION = "Alpha 1.0.42"
+VITE_APP_VERSION = "Alpha 1.0.46"
 VITE_APP_LANCADO = "N"
 VITE_SECRET_KEY = "f130c8fcf0694707421eecda9d04aafffdfb9948aa2f412edf47a733266f31921c31c56feb28beebbeb5c144637781b2f41042f5561bd5175cd503cdf8f6ef70"
