+++ Frontend/package.json
@@ -1,7 +1,7 @@
 {
   "name": "manutencao",
   "private": true,
-  "version": "alpha1.0.42",
+  "version": "alpha1.0.46",
   "type": "module",
   "scripts": {
     "dev": "vite",
