+++ Frontend/src/components/default/DefaultButton/interface/IDefaultButtonProps.tsx
@@ -1,0 +1,9 @@
+interface IDefaultButtonProps {
+  label: string;
+  severity?: string;
+  className?: string;
+  fullWidth?: boolean;
+  onClick?: React.MouseEventHandler<HTMLButtonElement>;
+}
+
+export default IDefaultButtonProps;
