+++ Frontend/src/components/templates/MainContent/style/MainContent.css
@@ -4,5 +4,4 @@
   flex-direction: column;
   position: relative;
   padding: 0.5rem !important;
-  height: 500px;
 }
