+++ Frontend/src/pages/Usuarios.tsx
@@ -1,6 +1,6 @@
 import { useEffect } from "react";
-import UsuariosCard from "../components/pages/Usuarios/UsuariosCard";
 import alterarTitlePagina from "../utils/generic/alterarTitlePagina";
+import UsuariosCard from "../components/pages/Usuarios/UsuariosCard";
 
 const Usuarios = (): JSX.Element => {
   useEffect((): void => {
